sensor-simulation/requirements.txt
# sensor-simulation/requirements.txt
geopy==2.2.0
kafka-python==2.0.2
